SFDM-1678 сравнения типа сущности

+++ src/main/java/ru/beeline/fdmnotificationsmanagement/dto/CapabilityParentDTO.java
@@ -14,5 +14,5 @@ import java.util.List;
 @AllArgsConstructor
 @Builder
 public class CapabilityParentDTO {
-    private List<Long> parents;
+    private List<Integer> parents;
 }
